backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
datasets==3.5.0
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dspy==2.6.17
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json_repair==0.41.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langdetect==1.0.9
litellm==1.63.7
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.2
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.4
onnx==1.17.0
openai==1.61.0
opencv-python==4.11.0.86
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
protobuf==6.30.2
pyarrow==19.0.1
pyclipper==1.3.0.post6
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyMuPDF==1.25.5
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-doctr==0.11.0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
scipy==1.15.2
setuptools==78.1.0
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
sympy==1.13.1
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
